Use Azure CLI to update the VMSS rather than azure-vmss plugin
azure-vmss plugin has a bug that when it updates the VMSS, it clears
the network security group associated with the VMSS virtual machine
profile. As a result, the resulting VM's will not be accessible by
the public network as the frontend Azure Load Balancer Standard requires
that the VMSS in the backend has NSG associated.